module_3(fix): form and inputs refactoring
