--- src/components/ui/form/index.ts
@@ -1,1 +1,0 @@
-export * from "./form";
