+++ src/components/ui/index.ts
@@ -1,3 +1,4 @@
+export { Select } from "./select";
 export { Label } from "./label";
 export { Input } from "./input";
 export { Accordion } from "./accordion";
