+++ src/components/ui/input/input.module.scss
@@ -10,13 +10,10 @@
 
   &::after {
     position: absolute;
-    display: flex;
-    justify-content: center;
-    align-items: center;
     width: 24px;
     height: 24px;
     top: 13px;
-    right: 8px;
+    right: 16px;
   }
 }
 
@@ -38,4 +35,33 @@
 
 .input__error {
   color: colors.$red;
+}
+
+input[type="date"] {
+  position: relative;
+}
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  width: auto;
+  height: auto;
+  color: transparent;
+  background: transparent;
+}
+
+input[type="date"]::-webkit-inner-spin-button,
+input[type="date"]::-webkit-clear-button {
+  z-index: -10;
+}
+
+input[type="date"]::-webkit-input-placeholder {
+  z-index: 99;
+}
+
+input[type="date"]:placeholder {
+  z-index: 99;
 }
