+++ src/main.css
@@ -39,3 +39,13 @@ body {
   min-width: 320px;
   min-height: 100vh;
 }
+
+input[type="number"]::-webkit-inner-spin-button,
+input[type="number"]::-webkit-outer-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+
+input[type="number"] {
+  -moz-appearance: textfield;
+}
